+++ src/DeviceComponents/Deadbolt.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Text } from "react-konva";
 import { TEXT_COLOR, LINE_COLOR } from "../constants";
 import { getDeadboltLockStatus } from "../Logics/AttributeLogics";
-import { DeadboltDevice } from "../Models/Devices";
+import { DeadboltDevice } from "../Models/devices";
 import { CommonProps } from "./Common";
 import { useSvgImage } from "../hooks/useSvgImage";
 import { Image } from "react-konva";
