+++ src/DeviceComponents/LightSwitch.tsx
@@ -4,7 +4,7 @@ import { Text } from "react-konva";
 import { TEXT_COLOR } from "../constants";
 import { ImageLightBulb } from "../PhysicalComponent/ImageLightBulb";
 import { getLightOnOffAttribute } from "../Logics/AttributeLogics";
-import { LightSwitchDevice } from "../Models/Devices";
+import { LightSwitchDevice } from "../Models/devices";
 import { CommonProps } from "./Common";
 export interface LightSwitchOptions extends CommonProps {
   deviceData: LightSwitchDevice;
