+++ src/DeviceComponents/Tv.tsx
@@ -2,7 +2,7 @@ import { CommonProps } from "./Common";
 import React, { useState } from "react";
 import { useInterval } from "../hooks/useInterval";
 import { getPowerOnAttribute } from "../Logics/AttributeLogics";
-import { TvDevice } from "../Models/Devices";
+import { TvDevice } from "../Models/devices";
 import { Text, Arc } from "react-konva";
 import { TEXT_COLOR } from "../constants";
 export enum TvDirection {
