+++ src/Models/AllDevices.ts
@@ -1,4 +1,4 @@
-import { DeviceDataKind } from "./Devices";
+import { DeviceDataKind } from "./devices";
 import { Dimmer } from "../DeviceComponents/Dimmer";
 import { LightSwitch } from "../DeviceComponents/LightSwitch";
 import { Contact, ContactDirection } from "../DeviceComponents/Contact";
@@ -143,133 +143,12 @@ import { WEST_WALL, NORTH_WALL, TEXT_PADDING } from "../constants";
 //     label: "",
 //     name: "",
 //   },
-//   "67": {
-//     id: "67",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Outside Garage Door",
-//     textPosition: [WEST_WALL + 2, NORTH_WALL + 334],
-//     direction: ContactDirection.West,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "32": {
-//     id: "32",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Indoor Garage Door",
-//     textPosition: [WEST_WALL + 260, NORTH_WALL + 320],
-//     direction: ContactDirection.East,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "31": {
-//     id: "31",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Front Door",
-//     textPosition: [WEST_WALL + 270, NORTH_WALL + 353],
-//     direction: ContactDirection.South,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "27": {
-//     id: "27",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Master Closet Door",
-//     textPosition: [WEST_WALL + 181, NORTH_WALL + 260],
-//     direction: ContactDirection.West,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "38": {
-//     id: "38",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Alicia Bedroom Window",
-//     textPosition: [WEST_WALL + 44, NORTH_WALL],
-//     direction: ContactDirection.SlideRight,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
 //   "483": {
 //     id: "483",
 //     kind: "AIRPURIFIER",
 //     component: AirPurifier,
 //     note: "Alicia Air Purifier",
 //     textPosition: [WEST_WALL + 93, NORTH_WALL + 38],
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "40": {
-//     id: "41",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Playroom Bedroom Window",
-//     textPosition: [WEST_WALL + 298, NORTH_WALL],
-//     direction: ContactDirection.SlideRight,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "41": {
-//     id: "41",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Jacob Bedroom Window",
-//     textPosition: [WEST_WALL + 168, NORTH_WALL],
-//     direction: ContactDirection.SlideRight,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "227": {
-//     id: "227",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Living Room Left Window",
-//     textPosition: [WEST_WALL + 348, NORTH_WALL + 384],
-//     direction: ContactDirection.SlideRight,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "226": {
-//     id: "226",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Living Room Right Window",
-//     textPosition: [WEST_WALL + 438, NORTH_WALL + 384],
-//     direction: ContactDirection.SlideLeft,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "18": {
-//     id: "18",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Kitchen Door",
-//     textPosition: [WEST_WALL + 516, NORTH_WALL + 28],
-//     direction: ContactDirection.SlideDown,
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "68": {
-//     id: "68",
-//     kind: "CONTACT",
-//     component: Contact,
-//     note: "Master Bedroom Left Window",
-//     textPosition: [WEST_WALL + 2, NORTH_WALL + 250],
-//     direction: ContactDirection.SlideUp,
 //     attributes: {},
 //     label: "",
 //     name: "",
@@ -459,26 +338,6 @@ import { WEST_WALL, NORTH_WALL, TEXT_PADDING } from "../constants";
 //     name: "",
 //     wattThreashold: 50,
 //   },
-//   "321": {
-//     id: "321",
-//     kind: "DEADBOLT",
-//     component: Deadbolt,
-//     note: "Front Door Deadbolt",
-//     textPosition: [WEST_WALL + 272, NORTH_WALL + 400],
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
-//   "322": {
-//     id: "322",
-//     kind: "DEADBOLT",
-//     component: Deadbolt,
-//     note: "Garage Indoor Deadbolt",
-//     textPosition: [WEST_WALL + 225, NORTH_WALL + 345],
-//     attributes: {},
-//     label: "",
-//     name: "",
-//   },
 //   "196": {
 //     id: "196",
 //     kind: "SWITCH",
@@ -648,6 +507,26 @@ export const allDevices: { [deviceId: string]: DeviceDataKind } = {
     note: "Kitchen Door",
     textPosition: [WEST_WALL + 516 + TEXT_PADDING, NORTH_WALL + 28 + TEXT_PADDING],
     direction: ContactDirection.SlideDown,
+    attributes: {},
+    label: "",
+    name: "",
+  },
+  "321": {
+    id: "321",
+    kind: "DEADBOLT",
+    component: Deadbolt,
+    note: "Front Door Deadbolt",
+    textPosition: [WEST_WALL + 272, NORTH_WALL + 400],
+    attributes: {},
+    label: "",
+    name: "",
+  },
+  "322": {
+    id: "322",
+    kind: "DEADBOLT",
+    component: Deadbolt,
+    note: "Garage Indoor Deadbolt",
+    textPosition: [WEST_WALL + 225, NORTH_WALL + 335],
     attributes: {},
     label: "",
     name: "",
